Title: 최솟값 3개, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240118/최솟값 3개/top-3-smallest-number.cpp
@@ -6,10 +6,10 @@ using namespace std;
 
 int main() {
 	vector<int> v;
-	int n, input;
+	long long n, input;
 	cin >> n;
 
-	for (int i = 0; i < n; i++)
+	for (long long i = 0; i < n; i++)
 	{
 		cin >> input;
 		v.push_back(input);
